Made the app to follow package structure 🤗
