+++ jesswinw@ieee.org/run.py
@@ -1,0 +1,4 @@
+from bloggy import app
+
+if __name__ == "__main__":
+    app.run(debug=True)
